Logic refactor and improvement
Target of the application is to charge a battery => The input current of the battery pack is the most important variable of the system. We thus try to follow the desired intensity with the buck boost SMPS. This intensity is given by the desired charge current for the battery pack + output current of the battery pack. This system seems to behave better than the previous MPPT algorithm. This update also refactors the code to be a bit more readable.

Pending further readability improvements (may delete unused sections)